requirements.txt
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.49.0
idna==3.6
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
tzdata==2024.1
urllib3==2.2.0
wcwidth==0.2.13
